+++ extensions/vscode/src/views/homeView.ts
@@ -1,4 +1,4 @@
-// Copyright (C) 2024 by Posit Software, PBC.
+// Copyright (C) 2025 by Posit Software, PBC.
 
 import path from "path";
 import debounce from "debounce";
@@ -108,6 +108,7 @@ import {
   isConnectCloudProduct,
   isConnectProduct,
 } from "src/utils/multiStepHelpers";
+import { recordAddConnectCloudUrlParams } from "src/utils/connectCloudHelpers";
 
 enum HomeViewInitialized {
   initialized = "initialized",
@@ -1912,7 +1913,10 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
         const response = await api.contentRecords.getAll(entrypointDir, {
           recursive: false,
         });
-        response.data.forEach((cfg) => {
+        const contentRecords = response.data.map((record) =>
+          recordAddConnectCloudUrlParams(record, env.appName),
+        );
+        contentRecords.forEach((cfg) => {
           if (!isContentRecordError(cfg)) {
             contentRecordList.push(cfg);
           }