+++ CHANGELOG.md
@@ -5,12 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.20.0]
+
+### Fixed
+
+- Fixed an issue deploying content to free Posit Connect Cloud accounts with incorrect  
+  access settings. (#3005)
+
+## [1.19.1]
 
 ### Fixed
 
 - Fixed the Python and R Packages views incorrectly stating the default package
   files were missing even when they were present (#2882, #2884)
+- Fixed an issue where the package file checkbox in the Project Files view
+  would occasionally be unchecked even when the file was included (#2793)
+
+### Added
+
+- Added UTM parameters for Connect Cloud links to see who is coming from
+  Publisher (#2969)
+- R dependencies are automatically detected when a project does not include an
+  `renv.lock` file.
+- Added a celebratory message when credential token is connected (#2901)
+
+### Changed
+
+- R manifest package generation now uses only `renv.lock` file by default. To use the
+  local `renv` library instead, set `[r].packages_from_library = true` in the
+  configuration.
 
 ## [1.19.0]